+++ pro_tes/celery_worker.py
@@ -1,8 +1,8 @@
 """Entry point for Celery workers."""
 
 from pro_tes.config.app_config import parse_app_config
-from pro_tes.factories.celery_app import create_celery_app
-from pro_tes.factories.connexion_app import create_connexion_app
+from foca.factories.celery_app import create_celery_app
+from foca.factories.connexion_app import create_connexion_app
 
 
 # Parse app configuration
