--- pro_tes/config/log_config.yaml
@@ -1,33 +1,0 @@
-version: 1
-
-disable_existing_loggers: False
-
-formatters:
-    standard:
-        class: logging.Formatter
-        style: "{"
-        format: "[{asctime}: {levelname:<8} {module:<18}] {message}"
-
-    long:
-        class: logging.Formatter
-        style: "{"
-        format: "[{asctime}: {levelname:<8}] {message} [{name}]"
-
-    # OTHER FORMATS
-        #format: "{message}"
-        #format: "[{asctime}] [{levelname:^8}] {message} ({name})"
-        #format: "{asctime}-{levelno:^2}-{name}-{module}-{funcName}: {message}"
-        #format: "[{asctime}: {levelname:}/{name:<36}] {message}"
-        #format: "[{asctime}] [{levelname:^8}] [{name}] {message} ({pathname}:{funcName})"
-        #datefmt: "%y-%m-%d %H:%M:%S"
-
-handlers:
-    console:
-        class: logging.StreamHandler
-        level: DEBUG
-        formatter: long
-        stream: ext://sys.stderr
-
-root:
-    level: INFO
-    handlers: [console]
