--- pro_tes/wsgi.py
@@ -1,3 +1,0 @@
-from pro_tes.app import run_server
-
-app, config = run_server()
